Refactor AI schema imports and enhance campaign strategy structure
- Updated import paths for various AI services and schemas to reflect the new directory structure, improving organization and clarity.
- Introduced new schemas for contact extraction and campaign strategy, enhancing data validation and integrity.
- Removed obsolete schemas related to contacts and smart filters, streamlining the codebase.

These changes aim to improve maintainability and ensure a clearer structure for AI-related functionalities.
