+++ server/src/modules/ai/schemas/contactExtraction/contactExtractionSchema.ts
