+++ server/src/modules/ai/schemas/siteAnalysis/reportOutputSchema.ts
