--- server/src/modules/ai/schemas/smartFilterSchemas.ts
@@ -1,7 +1,0 @@
-import { z } from 'zod';
-
-export const smartFilterSiteMapSchema = z.object({
-  urls: z.array(z.string()).describe('The filtered list of URLs'),
-});
-
-export type SmartFilterSiteMapOutput = z.infer<typeof smartFilterSiteMapSchema>;
