+++ server/src/types/timeout-transition.types.ts
@@ -1,4 +1,4 @@
-import type { CampaignPlanOutput } from '../modules/ai/schemas/contactCampaignStrategySchema';
+import type { CampaignPlanOutput } from '../modules/ai/schemas/contactStrategy/contactCampaignStrategySchema';
 import type { NextActionResult } from './campaign-transition.types';
 
 export interface ProcessTimeoutTransitionParams {
